board 작업

+++ src/main/java/himedia/myportal/services/BoardService.java
@@ -9,5 +9,5 @@ public interface BoardService {
 	public BoardVo getContent(Integer no);
 	public boolean write(BoardVo boardVo);
 	public boolean update(BoardVo boardVo);
-	public boolean deleteByNoAndUserNo(Integer no, Long userNo);
+	public boolean deleteByNoAndUserNo(Integer no, Integer userNo);
 }
